feat(fr): jolt-riscv per-instruction structs for FR coprocessor
Adds typed per-instruction structs under crates/jolt-riscv/src/instructions/field/
mirroring the existing i/m/a/virt layout:
- FieldOp        — funct3 selects FMUL/FADD/FSUB/FINV at runtime (static flags
                   left empty; this work will wire funct3-dispatched IsFieldMul/
                   Add/Sub/Inv from the witness-gen path)
- FieldAssertEq  — circuit flag IsFieldAssertEq
- FieldMov       — circuit flag IsFieldMov
- FieldSLL64     — circuit flag IsFieldSLL64
- FieldSLL128    — circuit flag IsFieldSLL128
- FieldSLL192    — circuit flag IsFieldSLL192

Wires each new struct into:
- instructions/mod.rs re-exports
- LookupInstruction enum + TryFrom<NormalizedInstruction> dispatch
- impl_jolt_instructions_flags! Flags fan-out

Widens CircuitFlagSet from u16 → u32 to hold 23 flags (was 14). Downstream
.bits() consumers in jolt-core/jolt-kernels work as-is since they only use
`bits() & (1 << index)`.

Validated: muldiv e2e (host) + 12 jolt-riscv tests + clippy clean.

+++ crates/jolt-riscv/src/flags.rs
@@ -94,21 +94,21 @@ pub const NUM_INSTRUCTION_FLAGS: usize = InstructionFlags::COUNT;
 
 /// Packed bitfield of [`CircuitFlags`].
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash)]
-pub struct CircuitFlagSet(u16);
+pub struct CircuitFlagSet(u32);
 
 impl CircuitFlagSet {
     #[inline]
     pub fn set(self, flag: CircuitFlags) -> Self {
-        Self(self.0 | (1 << flag as u16))
+        Self(self.0 | (1 << flag as u32))
     }
 
     #[inline]
     pub fn get(self, flag: CircuitFlags) -> bool {
-        self.0 & (1 << flag as u16) != 0
+        self.0 & (1 << flag as u32) != 0
     }
 
     #[inline]
-    pub fn bits(self) -> u16 {
+    pub fn bits(self) -> u32 {
         self.0
     }
 }
